+++ standingsrequests/app_settings.py
@@ -2,21 +2,6 @@ from django.conf import settings
 from .utils import clean_setting
 
 
-# id of character to use for updating alliance contacts - needs to be set
-STANDINGS_API_CHARID = clean_setting("STANDINGS_API_CHARID", None, required_type=int)
-
-STR_ALLIANCE_IDS = clean_setting("STR_ALLIANCE_IDS", None, required_type=list)
-
-STR_CORP_IDS = clean_setting("STR_CORP_IDS", None, required_type=list)
-
-# This is a map, where the key is the State the user is in
-# and the value is a list of required scopes to check
-SR_REQUIRED_SCOPES = getattr(
-    settings,
-    "SR_REQUIRED_SCOPES",
-    {"Member": ["publicData"], "Blue": [], "": []},  # no state
-)
-
 # switch to enable/disable ability to request standings for corporations
 SR_CORPORATIONS_ENABLED = clean_setting("SR_CORPORATIONS_ENABLED", True)
 
@@ -31,6 +16,20 @@ SR_OPERATION_MODE = clean_setting(
     "SR_OPERATION_MODE", "alliance", choices=["alliance", "corporation"]
 )
 
+# Grace period before a standing previously marked effective,
+# but that is no longer effective, will be reset
+SR_PREVIOUSLY_EFFECTIVE_GRACE_HOURS = clean_setting(
+    "SR_RESET_PREVIOUSLY_EFFECTIVE_MIN_HOURS", 2
+)
+
+# This is a map, where the key is the State the user is in
+# and the value is a list of required scopes to check
+SR_REQUIRED_SCOPES = getattr(
+    settings,
+    "SR_REQUIRED_SCOPES",
+    {"Member": ["publicData"], "Blue": [], "": []},  # no state
+)
+
 # Standings revocations will be considered stale and removed
 # from the local database after the configured days
 SR_REVOCATIONS_STALE_DAYS = clean_setting("SR_REVOCATIONS_STALE_DAYS", 7)
@@ -39,3 +38,16 @@ SR_REVOCATIONS_STALE_DAYS = clean_setting("SR_REVOCATIONS_STALE_DAYS", 7)
 # database after the configured hours.
 # The latest standings data will never be purged, no matter how old it is
 SR_STANDINGS_STALE_HOURS = clean_setting("SR_STANDINGS_STALE_HOURS", 48)
+
+# Max hours to wait for a standing to be effective after being marked actioned
+# Non effective standing requests will be reset when this timeout expires.
+SR_STANDING_TIMEOUT_HOURS = clean_setting("SR_STANDING_TIMEOUT_HOURS", 24)
+
+# id of character to use for updating alliance contacts - needs to be set
+STANDINGS_API_CHARID = clean_setting("STANDINGS_API_CHARID", None, required_type=int)
+
+# charaters belonging to these alliances are considered to be "in organization"
+STR_ALLIANCE_IDS = clean_setting("STR_ALLIANCE_IDS", None, required_type=list)
+
+# charaters belonging to these corporations are considered to be "in organization"
+STR_CORP_IDS = clean_setting("STR_CORP_IDS", None, required_type=list)